app/requirements.txt
flask
flask-pymongo
flask-jwt-extended
python-dotenv
bcrypt
cryptography
itsdangerous
python-multipart
pytest